requirements.txt
streamlit==1.30.0
pandas==2.1.4
sentence-transformers==2.2.2
umap-learn==0.5.5
plotly==5.18.0
scikit-learn==1.3.2
numpy==1.26.2
